not important

+++ src/training/learner/learner.py
@@ -79,7 +79,7 @@ class Learner:
                 loss value computed using batch's data
         """
         if self._params["reduce_dimensions"]:
-            batch["image"] = self._encode_image(batch["image"])
+            batch["image"] = self._encode_image(batch["image"].type(torch.float32))
         noise, noise_pred = self._forward(batch)
 
         # Loss backward
